Move tcp to separate file
